feat: change rating column type to unsignedTinyInteger in contacts table

+++ database/migrations/2025_06_11_142218_add_rating_to_contacts_table.php
@@ -9,7 +9,7 @@ return new class() extends Migration
     public function up(): void
     {
         Schema::table('contacts', function (Blueprint $table): void {
-            $table->unsignedInteger('rating')->default(0)->after('debtor_number');
+            $table->unsignedTinyInteger('rating')->default(0)->after('debtor_number');
         });
     }
